+++ src/main/java/duke/command/ByeCommand.java
@@ -1,3 +1,9 @@
+package duke.command;
+
+import duke.task.TaskList;
+import duke.Ui;
+import duke.Storage;
+
 public class ByeCommand extends Command {
 
     @Override
