+++ src/main/java/duke/command/Command.java
@@ -1,4 +1,8 @@
-import java.io.IOException;
+package duke.command;
+
+import duke.task.TaskList;
+import duke.Ui;
+import duke.Storage;
 
 public abstract class Command {
     public abstract void execute(TaskList taskList, Ui ui, Storage storage);
