+++ src/main/java/duke/command/UnmarkCommand.java
@@ -1,16 +1,23 @@
-public class DeleteCommand extends Command {
+package duke.command;
+
+import duke.task.Task;
+import duke.task.TaskList;
+import duke.Ui;
+import duke.Storage;
+
+public class UnmarkCommand extends Command {
 
     private final int index;
 
-    public DeleteCommand(int index) {
+    public UnmarkCommand(int index) {
         this.index = index;
     }
 
     @Override
     public void execute(TaskList taskList, Ui ui, Storage storage) {
         Task task = taskList.getTask(index);
-        taskList.deleteTask(index);
-        ui.showTaskDeleted(task, taskList);
+        task.markAsNotDone();
+        ui.showTaskUnmarked(task);
         storage.saveData(taskList, ui);
     }
