+++ src/main/java/duke/exception/DukeException.java
@@ -1,5 +1,6 @@
-public class DukeException extends Exception {
+package duke.exception;
 
+public class DukeException extends Exception {
     public DukeException(String message) {
         super(message);
     }
